+++ src/main/java/lotto/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.validator;
 
 import java.util.HashSet;
 import java.util.List;
